Field 90

+++ SharpMT942/Statement.cs
@@ -7,13 +7,13 @@ namespace SharpMT942
         public DateTime ValueDate { get; }
         public DateTime EntryDate { get; }
         public TransactionDirection DebitOrCredit { get; }
-        public double Amount { get; }
+        public decimal Amount { get; }
         public TransactionType TransactionType { get; }
         public string CustomerReference { get; }
         public string BankReference { get; }
         public string SupplementaryDetails { get; }
 
-        public Statement (DateTime valueDate, DateTime entryDate, TransactionDirection transactionDirection, double amount, TransactionType transactionType, string customerReference, string bankReference, string supplementaryDetails)
+        public Statement (DateTime valueDate, DateTime entryDate, TransactionDirection transactionDirection, decimal amount, TransactionType transactionType, string customerReference, string bankReference, string supplementaryDetails)
         {
             ValueDate = valueDate;
             EntryDate = entryDate;
